First iteration of shooting and damaging works
Shooting now works during the first shot. Damage is also applied to the shootable tiles current occupant, and if the occupants health goes to 0 or under they are destroyed. What happens after a tank takes a shot (or tries to shoot) still needs to be addressed (unselect etc.), game bugs out after the first shot.